+++ src/components/__tests__/CvssCalculatorOverlayed.spec.ts
@@ -1,35 +1,54 @@
-import { mount } from '@vue/test-utils';
+import { type Component } from 'vue';
+
+import sampleFlawFull from '@test-fixtures/sampleFlawFull.json';
+import { VueWrapper, mount } from '@vue/test-utils';
 import { describe, it, expect } from 'vitest';
 
-import CvssCalculatorOverlayed from '@/components/CvssCalculator/CvssCalculatorOverlayed.vue';
+import { useCvssScores, validateCvssVector } from '@/composables/useCvssScores';
 
-function activateFactorButton(subject: any, factor: string, value: string) {
+import { importActual } from '@/__tests__/helpers';
+
+vi.mock('@/composables/useCvssScores');
+
+async function activateFactorButton(subject: any, factor: string, value: string) {
   const factorButtonGroup = subject.findAll('.btn-group-vertical')
     .filter((node: { text: () => string | string[] }) => node.text().includes(factor))[0];
   const factorButton = factorButtonGroup
     .findAll('.btn')
     .filter((node: { text: () => string }) => node.text() === value)[0];
-  factorButton.trigger('click');
+  await factorButton.trigger('click');
 }
 
-function activateMultipleFactorButtons(subject: any, factorValuePairs: [string, string][]) {
+async function activateMultipleFactorButtons(subject: any, factorValuePairs: [string, string][]) {
   for (const [factor, value] of factorValuePairs) {
-    activateFactorButton(subject, factor, value);
+    await activateFactorButton(subject, factor, value);
   }
 }
 
 describe('cvssCalculatorOverlayed', () => {
-  let subject: any;
+  type Subject = VueWrapper<Component>;
+  let subject: Subject;
+  beforeEach(async () => {
+    const {
+      useCvssScores: _useCvssScores,
+      validateCvssVector: _validateCvssVector,
+    } = await importActual('@/composables/useCvssScores');
+    vi.mocked(useCvssScores).mockImplementation(_useCvssScores);
+    vi.mocked(validateCvssVector).mockImplementation(_validateCvssVector);
 
-  beforeEach(() => {
+    const importedComponent = await import('@/components/CvssCalculator/CvssCalculatorOverlayed.vue');
+    const CvssCalculatorOverlayed = importedComponent.default;
     subject = mount(CvssCalculatorOverlayed, {
       props: {
-        'cvssScore': null,
-        'onUpdate:cvssScore': (e: any) => subject.setProps({ cvssScore: e }),
-        'cvssVector': '',
-        'onUpdate:cvssVector': (e: any) => subject.setProps({ cvssVector: e }),
+        affect: sampleFlawFull.affects[0],
       },
     });
+  });
+
+  afterEach(() => {
+    vi.resetAllMocks();
+    vi.resetModules();
+    subject?.unmount();
   });
 
   it('mounts and renders correctly', async () => {
@@ -66,15 +85,12 @@ describe('cvssCalculatorOverlayed', () => {
     ];
     await activateMultipleFactorButtons(subject, factorValuePairs);
     const inputVectorValue = subject.find('.vector-input');
-    expect(inputVectorValue.classes().includes('is-invalid')).toBe(false);
+    expect(inputVectorValue.classes().includes('is-p')).toBe(false);
   });
 
   it('shows validation on incomplete vector', async () => {
-    const factorValuePairs: [string, string][] = [
-      ['Availability', 'High'],
-    ];
-    await activateMultipleFactorButtons(subject, factorValuePairs);
     const inputVectorValue = subject.find('.vector-input');
+    await activateFactorButton(subject, 'Availability', 'High');
     expect(inputVectorValue.classes().includes('is-invalid')).toBe(true);
   });
 });