+++ src/composables/__tests__/useFlawModel.spec.ts
@@ -27,6 +27,7 @@ vi.mock('@/services/FlawService', () => ({
   postFlaw: vi.fn().mockResolvedValue({}),
   putFlaw: vi.fn().mockResolvedValue({}),
   putFlawCvssScores: vi.fn().mockResolvedValue({}),
+  deleteFlawCvssScores: vi.fn().mockResolvedValue({}),
 }));
 
 describe('useFlawModel', () => {
@@ -119,11 +120,10 @@ describe('useFlawModel', () => {
     });
 
     it('should prevent saving if CVSS scores are invalid', async () => {
-      const { flaw } = useFlaw();
-      flaw.value = sampleFlawFull as ZodFlawType;
-      const { flawRhCvss3, updateFlaw } = mountFlawModel();
-
-      flawRhCvss3.value.vector = 'invalid';
+      const { setFlaw } = useFlaw();
+      setFlaw(sampleFlawFull as ZodFlawType);
+      const { updateFlaw, updateVector } = mountFlawModel();
+      updateVector('not valid');
       await flushPromises();
       updateFlaw();