+++ CHANGELOG.md
@@ -17,6 +17,9 @@
 * Improve loading performance by splitting up flaw and affect requests  (`OSIDB-4266`)
 * Don't update Flaw when only affects are modified (`OSIDB-4270`)
 
+### Added
+* CVSS4 support, disabled until adoption (`OSIDB-2363`)
+
 ## [2025.6.0]
 ### Added
 * Add an unified view for all comment types  (`OSIDB-3467`)